+++ src/app/(auth)/_components/user-menu.tsx
@@ -2,13 +2,11 @@
 
 import { FaUserGroup } from 'react-icons/fa6'
 import { IoExitOutline } from 'react-icons/io5'
-import { FaUserShield } from 'react-icons/fa'
-import type { User } from '@supabase/supabase-js'
+import { FaUser, FaUserShield } from 'react-icons/fa'
 import { useRouter } from 'next/navigation'
 import { toast } from 'sonner'
 
 import { signOut } from '@/supabase/auth.service'
-import { api } from '@/trpc/react'
 import { AUTHOR_EMAIL } from '@/data/constants'
 
 import { Avatar, AvatarFallback, AvatarImage } from '@/ui/avatar'
@@ -22,6 +20,7 @@ import {
   DropdownMenuSeparator,
   DropdownMenuTrigger,
 } from '@/ui/dropdown-menu'
+import { AuthLeetcoder } from '@/server/routers/auth'
 
 const MenuItem = ({ icon, label, onClick }: { icon: React.ReactNode; label: string; onClick: () => Promise<void> }) => (
   <DropdownMenuItem onClick={onClick}>
@@ -30,20 +29,11 @@ const MenuItem = ({ icon, label, onClick }: { icon: React.ReactNode; label: stri
   </DropdownMenuItem>
 )
 
-const renderAvatar = (user: User) => (
-  <Avatar>
-    <AvatarImage
-      src={user?.user_metadata?.avatar_url || ''}
-      alt={user?.user_metadata?.full_name || ''}
-    />
-    <AvatarFallback>{user?.user_metadata?.full_name?.substring(0, 2)}</AvatarFallback>
-  </Avatar>
-)
-
-export const UserMenu = ({ user }: { user: User }) => {
+export const UserMenu = ({ user }: { user: AuthLeetcoder }) => {
   const router = useRouter()
-  const { data: leetcoder } = api.leetcoders.getLeetcoderById.useQuery({ id: user.id })
-  const isAdmin = leetcoder?.email === AUTHOR_EMAIL
+
+  const isAdmin = user.email === AUTHOR_EMAIL
+  const isMember = user.leetcoder?.status === 'APPROVED' || user.leetcoder?.status === 'PENDING'
 
   const handleLogout = async () => {
     const { error } = await signOut()
@@ -58,9 +48,13 @@ export const UserMenu = ({ user }: { user: User }) => {
     router.push('/dashboard')
   }
 
+  const navigateToProfile = async () => {
+    router.push('/profile')
+  }
+
   const navigateToGroup = async () => {
-    if (leetcoder?.group_no) {
-      router.push(`/group/${leetcoder.group_no}`)
+    if (user.leetcoder?.group_no) {
+      router.push(`/group/${user.leetcoder.group_no}`)
     }
   }
 
@@ -71,14 +65,47 @@ export const UserMenu = ({ user }: { user: User }) => {
           variant="ghost"
           className="h-auto cursor-pointer rounded-full p-0 hover:bg-transparent"
         >
-          {renderAvatar(user)}
+          <Avatar>
+            <AvatarImage
+              src={user?.leetcoder?.avatar || ''}
+              alt={user?.leetcoder?.name || ''}
+            />
+            <AvatarFallback>{user?.user_metadata?.full_name?.substring(0, 2)}</AvatarFallback>
+          </Avatar>
         </Button>
       </DropdownMenuTrigger>
       <DropdownMenuContent className="max-w-64">
         <DropdownMenuLabel className="flex min-w-0 flex-col">
           <span className="text-muted-foreground truncate text-xs font-normal">{user?.email}</span>
         </DropdownMenuLabel>
-        <DropdownMenuSeparator />
+
+        {isMember && (
+          <>
+            <DropdownMenuSeparator />
+
+            <MenuItem
+              icon={
+                <FaUserGroup
+                  size={16}
+                  className="opacity-60"
+                />
+              }
+              label="My Group"
+              onClick={navigateToGroup}
+            />
+
+            <MenuItem
+              icon={
+                <FaUser
+                  size={16}
+                  className="opacity-60"
+                />
+              }
+              label="Profile"
+              onClick={navigateToProfile}
+            />
+          </>
+        )}
 
         {isAdmin && (
           <MenuItem
@@ -92,17 +119,6 @@ export const UserMenu = ({ user }: { user: User }) => {
             onClick={navigateToDashboard}
           />
         )}
-
-        <MenuItem
-          icon={
-            <FaUserGroup
-              size={16}
-              className="opacity-60"
-            />
-          }
-          label="My Group"
-          onClick={navigateToGroup}
-        />
 
         <DropdownMenuGroup />
         <DropdownMenuSeparator />