+++ src/app/group/[groupId]/page.tsx
@@ -1,33 +1,21 @@
-// import { api } from '@/trpc/server'
-// import { generateTableData } from '@/utils/generateTableData'
-// import { redis } from '@/config/redis'
-// import type { GroupData } from '@/types/tableRow.type'
-// import type { roadmap } from '@prisma/client'
-// import { NOT_STARTED_GROUPS } from '@/data/constants'
+import { api } from '@/trpc/server'
+import { generateTableData } from '@/utils/generateTableData'
+import { NOT_STARTED_GROUPS } from '@/data/constants'
 
-// import { TrackTable } from '@/app/group/_components/track-table'
-// import { ConfettiFireworks } from '@/app/group/_components/confetti-fireworks'
-// import { NotStarted } from '@/app/group/_components/not-started'
-import { Maintenance } from '@/app/_components/maintenance'
+import { TrackTable } from '@/app/group/_components/track-table'
+import { ConfettiFireworks } from '@/app/group/_components/confetti-fireworks'
+import { NotStarted } from '@/app/group/_components/not-started'
 
-// type CachedData = {
-//   groupData: GroupData
-//   roadmap: roadmap[]
-// }
-
-// const Page = async ({ params }: { params: Promise<{ groupId: string }> }) => {
-const Page = async () => {
-  return <Maintenance />
-
-  // const { groupId } = await params
+const Page = async ({ params }: { params: Promise<{ groupId: string }> }) => {
+  const { groupId } = await params
 
   // for not started groups
-  // if (NOT_STARTED_GROUPS.includes(+groupId)) return <NotStarted />
+  if (NOT_STARTED_GROUPS.includes(+groupId)) return <NotStarted />
 
-  // const user = await api.auth.getUser()
-  // console.log('user?.user_metadata', user?.user_metadata)
-  // const groupData = await api.groups.getGroupTableData({ group_no: groupId })
-  // const roadmap = await api.roadmap.getGroupProblems(groupData?.group_progress || [])
+  const user = await api.auth.getUser()
+  console.log('user?.user_metadata', user?.user_metadata)
+  const groupData = await api.groups.getGroupTableData({ group_no: groupId })
+  const roadmap = await api.roadmap.getGroupProblems(groupData?.group_progress || [])
 
   // const cacheKey = `group:${groupId}:data`
   // let groupData: GroupData | null = null
@@ -49,26 +37,26 @@ const Page = async () => {
   //   }
   // }
 
-  // if (!groupData || !roadmap || roadmap.length === 0) return <NotStarted />
+  if (!groupData || !roadmap || roadmap.length === 0) return <NotStarted />
 
-  // const tableData = generateTableData({
-  //   group_no: groupData.group_no,
-  //   submission: groupData.submissions,
-  //   roadmap,
-  //   group_progress: groupData.group_progress,
-  // })
+  const tableData = generateTableData({
+    group_no: groupData.group_no,
+    submission: groupData.submissions,
+    roadmap,
+    group_progress: groupData.group_progress,
+  })
 
-  // return (
-  //   <>
-  //     <TrackTable
-  //       leetcoders={groupData.leetcoders}
-  //       tableData={tableData}
-  //       groupId={groupId}
-  //     />
+  return (
+    <>
+      <TrackTable
+        leetcoders={groupData.leetcoders}
+        tableData={tableData}
+        groupId={groupId}
+      />
 
-  //     <ConfettiFireworks />
-  //   </>
-  // )
+      <ConfettiFireworks />
+    </>
+  )
 }
 
 export default Page