+++ src/types/leetcoders.type.ts
@@ -1,4 +1,6 @@
 import type { submissions, leetcoders, groups, roadmap } from '@prisma/client'
+import type { submissions } from '@prisma/client'
+import { z } from 'zod'
 
 export type LeetcoderWithSubmissions = {
   id: string
@@ -9,8 +11,24 @@ export type LeetcoderWithSubmissions = {
   submissions: Pick<submissions, 'problem_id'>[]
 }
 
+
 // Definition for LeetcoderWithProblem
 export type LeetcoderWithProblem = leetcoders & {
   group: groups
   problemDetails: Pick<roadmap, 'problem_slug' | 'difficulty' | 'topic'>
 }
+
+export const updateLeetcoderSchema = z.object({
+  id: z.string(),
+  gh_username: z.string().optional(),
+  x_username: z.string().optional(),
+  li_username: z.string().optional(),
+  group_no: z
+    .string()
+    .transform((val) => Number(val))
+    .optional(),
+  website: z.string().optional(),
+  is_visible: z.boolean().optional(),
+})
+
+export type UpdateLeetcoderInput = z.infer<typeof updateLeetcoderSchema>